+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Fixed
 - Blacked code.
 - Bugs crashing query when probe has 1 or 2 oligos only.
+- Bugs crashing query when probe set has 1 or 2 probes only.
 
 ## Removed
 - `web` module and all network-based checks.